Add more haddock docs for AST.Lexical.
Fill in the haddock docs for the rest of the types, and expand on the
module comment. I also decided what the policy on preserving the
original input text is: unparsing will not be guaranteed to reproduce
Unicode escapes as they were in the original source text.

I've also described some requirements for the behaviour of lexers.

The next steps should be to define a lexer and a way to unparse
a stream of JInputElements or JTokens, and build some unparse-parse
QuickCheck tests.