Add some simple character-counting benchmarks.
Counting characters is going to be a lot lighter-weight than lexing,
so it represents an upper bound on how fast a lexer can be.

I've written a small character-counting program, with a few slightly
different approaches to the counting, and a Criterion benchmark. We
can count the characters in /usr/share/dict/words (English Canadian
locale with UTF-8 default encoding) in between 130us and 135us.

This is three or four times slower than the unix 'wc' program. I
haven't pin-pointed why yet, even with profiling. But making this
much faster isn't really top priority, since lexing will probably
be a lot slower than character counting.