Add a first cut at Unicode escapes.
Add Language.Java.Lexer.UnicodeEscapes, with functions for converting
Unicode escapes, as described in the JLS, to and from Unicode
characters.

This code isn't as simple as it could be, because I got a bit
obsessed with making it fast. After a bunch of criterion-assisted
testing, I got the 'escapeToChar' function down to about 20ns
on my machine using GHC 6.10. However, with GHC 6.12, the time
is back up to about 50ns. I'll add the benchmarks in a separate
commit, and try to figure out why the benchmark time has changed.

I've also removed some unnecessary dependencies from the
cabal file, and switched to version 4 of base.